in progress
working on fixing captcha, and updating to v3
