+++ globals.d.ts
@@ -1,0 +1,9 @@
+declare namespace NodeJS {
+	interface ProcessEnv {
+		NEXT_PUBLIC_RECAPTCHA_SITE_KEY: string;
+		RECAPTCHA_SECRET_KEY: string;
+		RECAPTCHA_SECRET_KEY: string;
+		EMAIL_TO: string;
+		MAIL_USER: string;
+	}
+}
