+++ package.json
@@ -46,6 +46,7 @@
 		"react": "19.0.0",
 		"react-dom": "19.0.0",
 		"react-google-recaptcha": "^3.1.0",
+		"react-google-recaptcha-v3": "^1.10.1",
 		"react-hook-form": "^7.54.2",
 		"sharp": "^0.33.5",
 		"tailwind-merge": "^2.6.0",
